+++ layers/Builder.py
@@ -1,4 +1,5 @@
 import os.path
+from pprint import pprint
 from typing import Generic, TypeVar
 
 from engine.layer_0 import set_solc_version_by_content
@@ -14,15 +15,20 @@ class BaseBuilder(Generic[Buildable]):
     def __init__(self):
         self.loader = Loader()
 
-    def build(self, code) -> Buildable:
-        raise NotImplementedError("Method not implemented")
+    def build(self, _code) -> str:
+        _path = self.loader.cache_content(_code, "sol")
+        if not os.path.exists(_path):
+            _path = self.loader.cache_content(_code, "sol")
+        self.loader.format(_path)
+        return _path
 
 
 class ContractBuilder(BaseBuilder[Contract]):
     def __init__(self):
         super().__init__()
 
-    def build(self, _path) -> Contract:
+    def build(self, _code) -> Contract:
+        _path = super().build(_code)
         _code = self.loader.read_code(_path)
         _version = set_solc_version_by_content(_code)
         return Contract(
@@ -38,7 +44,12 @@ class FunctionBuilder(BaseBuilder[Function]):
     def __init__(self):
         super().__init__()
 
-    def build(self, code) -> list[Function]:
+    def build(self, _code) -> list[Function]:
+        _path = super().build(_code)
+        return FunctionBuilder.mock()
+
+    @staticmethod
+    def mock() -> list[Function]:
         return [
             Function(
                 name="setValue",
@@ -49,15 +60,15 @@ class FunctionBuilder(BaseBuilder[Function]):
                 mutability="",
                 returns=[],
                 body="""
-        require(_value > 0, "Value must be greater than zero");
-        assert(value != _value);
-        require(isEven(_value), "Value must be even");
-        if (_value < 10) {
-            revert ValueTooLow({provided: _value, required: 10});
-        }
-        revert();
-        revert("This is a revert without a custom error");
-        value = _value;""".strip(),
+                    require(_value > 0, "Value must be greater than zero");
+                    assert(value != _value);
+                    require(isEven(_value), "Value must be even");
+                    if (_value < 10) {
+                        revert ValueTooLow({provided: _value, required: 10});
+                    }
+                    revert();
+                    revert("This is a revert without a custom error");
+                    value = _value;""".strip(),
                 _has_low_level_call=True
             ),
             Function(
@@ -78,7 +89,22 @@ class VariableBuilder(BaseBuilder[Variable]):
     def __init__(self):
         super().__init__()
 
-    def build(self, code) -> list[Variable]:
+    def build(self, _code) -> list[Variable]:
+        _path = super().build(_code)
+        # 1. Loader should format the source code
+        # 2. Parser should extract the variables
+        import parser
+        res = parser.layer_2.get_variables(_path)
+
+        # 3. Builder should build the variables
+        print()
+        print(res.keys())
+        pprint(res)
+
+        return VariableBuilder.mock()
+
+    @staticmethod
+    def mock() -> list[Variable]:
         return [
             Variable(
                 name="value",
@@ -126,117 +152,3 @@ class VariableBuilder(BaseBuilder[Variable]):
                 mutability=Mutability.MUTABLE
             )
         ]
-
-
-_target_code: str = """// SPDX-License-Identifier: MIT
-            pragma solidity ^0.8.0;contract ComplexChecks is A, B { using aa for *; uint256 public value; error ValueTooLow(uint256 provided, uint256 required);constructor(uint256 _initialValue) {      value = _initialValue;}function setValue(uint256 _value) public aa b(this) override {require(_value > 0, "Value must be greater than zero"); assert(value != _value); require(isEven(_value), "Value must be even"); if (_value < 10) {revert ValueTooLow({provided: _value, required: 10});} revert();     revert("This is a revert without a custom error"); value = _value;}function isEven(uint256 _value) internal pure returns (bool) {return _value % 2 == 0;}}"""
-_cached = "49cd11d435ac68a380b5c14dd54b9720fa27bbd4250adb9b1136c7fa4673e784.sol"
-_cached_path = BaseBuilder().loader.as_abs_path(os.path.join("out", _cached))
-
-
-def test_contract_builder():
-    builder = ContractBuilder()
-    _path = _cached_path
-    if not os.path.exists(_path):
-        _path = builder.loader.cache_content(_target_code, "sol")
-    assert builder.build(_path) == Contract(
-        target_code=_target_code,
-        version="0.8.0",
-        name="ComplexChecks",
-        inheritance=["A", "B"],
-        library=["aa"]
-    )
-
-
-def test_function_builder():
-    builder = FunctionBuilder()
-    _path = _cached_path
-    if not os.path.exists(_path):
-        _path = builder.loader.cache_content(_target_code, "sol")
-    assert builder.build(_path) == [
-        Function(
-            name="setValue",
-            parameters=[],
-            visibility="public",
-            modifiers=["aa", "b(this)"],
-            is_override=True,
-            mutability="",
-            returns=[],
-            body="""
-        require(_value > 0, "Value must be greater than zero");
-        assert(value != _value);
-        require(isEven(_value), "Value must be even");
-        if (_value < 10) {
-            revert ValueTooLow({provided: _value, required: 10});
-        }
-        revert();
-        revert("This is a revert without a custom error");
-        value = _value;""".strip(),
-            _has_low_level_call=True
-        ),
-        Function(
-            name="isEven",
-            parameters=[],
-            visibility="internal",
-            modifiers=[],
-            is_override=False,
-            mutability="pure",
-            returns=[],
-            body="return _value % 2 == 0;",
-            _has_low_level_call=False
-        )
-    ]
-
-
-def test_variable_builder():
-    builder = VariableBuilder()
-    _path = _cached_path
-    if not os.path.exists(_path):
-        _path = builder.loader.cache_content(_target_code, "sol")
-    assert builder.build(_path) == [
-        Variable(
-            name="value",
-            signature="ComplexChecks",
-            type="uint256",
-            location=Location.STORAGE,
-            visibility=Visibility.PUBLIC,
-            scope=Scope.STORAGE,
-            mutability=Mutability.MUTABLE
-        ),
-        Variable(
-            name="_value",
-            signature="ComplexChecks:setValue",
-            type="bytes",
-            location=Location.CALLDATA,
-            visibility=Visibility.PUBLIC,
-            scope=Scope.ARGS,
-            mutability=Mutability.IMMUTABLE
-        ),
-        Variable(
-            name="result",
-            signature="ComplexChecks:setValue",
-            type="bytes",
-            location=Location.MEMORY,
-            visibility=Visibility.INTERNAL,
-            scope=Scope.RETURNS,
-            mutability=Mutability.MUTABLE
-        ),
-        Variable(
-            name="_value",
-            signature="ComplexChecks:isEven",
-            type="uint256",
-            location=Location.MEMORY,
-            visibility=Visibility.INTERNAL,
-            scope=Scope.ARGS,
-            mutability=Mutability.MUTABLE
-        ),
-        Variable(
-            name="result",
-            signature="ComplexChecks:isEven",
-            type="bool",
-            location=Location.MEMORY,
-            visibility=Visibility.INTERNAL,
-            scope=Scope.RETURNS,
-            mutability=Mutability.MUTABLE
-        )
-    ]
